Moved number functions to queries/num.py for now
